flask-server/requirements.txt
alabaster==0.7.10
astroid==1.5.3
attrdict==2.0.0
Babel==2.5.0
bleach==1.5.0
certifi==2018.8.24
cffi==1.10.0
chardet==3.0.4
click==6.7
colorama==0.3.9
cytoolz==0.9.0.1
decorator==4.1.2
dlib==19.15.0
docutils==0.14
entrypoints==0.2.3
eth-abi==1.2.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-typing==1.3.0
eth-utils==1.2.2
face-recognition==1.2.2
face-recognition-models==0.3.0
Flask==1.0.2
hexbytes==0.1.0
html5lib==0.9999999
idna==2.7
imageio==2.2.0
imagesize==0.7.1
imutils==0.4.6
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
isort==4.2.15
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.3.1
lru-dict==1.1.6
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.4.0
numpy==1.13.1
numpydoc==0.7.0
olefile==0.44
pandocfilters==1.4.2
parsimonious==0.8.0
path.py==10.3.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
psutil==5.2.2
pycodestyle==2.3.1
pycparser==2.18
pycryptodome==3.6.6
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.2
pypiwin32==223
python-dateutil==2.6.1
pytz==2017.2
pywin32==223
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.19.1
rlp==1.0.2
rope-py3k==0.9.4.post1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
socketIO-client-nexus==0.7.6
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.3
testpath==0.3.1
toolz==0.9.0
tornado==4.5.2
traitlets==4.3.2
typing==3.6.4
urllib3==1.23
wcwidth==0.1.7
web3==4.7.1
webencodings==0.5.1
websocket-client==0.53.0
websockets==6.0
Werkzeug==0.14.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
